% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mock_surface.R
\name{create_mock_surface}
\alias{create_mock_surface}
\title{create_mock_surface}
\usage{
create_mock_surface(
  raster.list,
  multiple.extents = F,
  resolution = list(x = 100, y = 100)
)
}
\value{
Logical
}
\description{
create_mock_surface
}
