% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_subgraph_centrality.R
\name{grab_subgraph_centrality}
\alias{grab_subgraph_centrality}
\title{grab_subgraph_centrality}
\usage{
grab_subgraph_centrality(pred.surf.comp, new.neighborhood = NULL)
}
\description{
grab_subgraph_centrality
}
