% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_finder.R
\name{neighbor_finder}
\alias{neighbor_finder}
\title{neighbor_finder}
\usage{
neighbor_finder(
  ssf.obj = m2,
  cell.data,
  neighbors.found,
  quantile = 0.99,
  cell.data.list = NULL,
  distance.override = NULL
)
}
\description{
neighbor_finder
}
