% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ssf_comparisons.R
\name{predict_ssf_comparisons}
\alias{predict_ssf_comparisons}
\title{predict_ssf_comparisons}
\usage{
predict_ssf_comparisons(ssf.obj = issf.fit, ssf.comparisons)
}
\description{
predict_ssf_comparisons
}
